refactor(port): update GetBeneficiaryAccountUseCase for Long ID
Función: Puerto de entrada para la recuperación de detalles de una cuenta de beneficiario.
Cambios: Se actualizó el parámetro ID de Integer a Long.

+++ src/main/java/org/walrex/application/port/input/GetBeneficiaryAccountUseCase.java
@@ -4,5 +4,5 @@ import io.smallrye.mutiny.Uni;
 import org.walrex.domain.model.BeneficiaryAccount;
 
 public interface GetBeneficiaryAccountUseCase {
-    Uni<BeneficiaryAccount> findById(Integer id);
+    Uni<BeneficiaryAccount> findById(Long id);
 }
